pyproject.toml
[tool.poetry]
name = "wheeler-memory"
version = "0.1.0"
description = "A standalone, biological-grade associative memory system for AI consciousness."
authors = ["Ralph Team <tristan@example.com>"]
readme = "README.md"
packages = [{include = "wheeler"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.0"
numpy = "^1.24.0"
aiosqlite = "^0.19.0"
click = "^8.1.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.0"
ruff = "^0.1.0"
mypy = "^1.0.0"
hypothesis = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=wheeler"
testpaths = [
    "tests",
]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
